move logic to features dir
